filterBestPokemon useMemo
