+++ src/components/Hero/useFetchFirebase.js
@@ -227,41 +227,3 @@ export const fetchPopularPokemon = () => {
         // };
     }, []);
 };
-
-export const filterPopularPokemon = () => {
-    const {
-        bestDamage,
-        popularPokemon,
-        storeAllPokemon,
-        handleFilterBestPokemon,
-    } = useHeroContext();
-    useEffect(() => {
-        const filterPokemon = storeAllPokemon.filter((allItem) => {
-            if (allItem === undefined) return;
-            return popularPokemon.some(
-                (popularItem) => allItem.enName === popularItem.enName
-            );
-        });
-
-        const bestPokemon = (filterPokemon ?? []).filter((pokemon) => {
-            let isReturn = false;
-            pokemon.Types.forEach((type) => {
-                bestDamage.forEach((best) => {
-                    if (best.zhName === type.zhName) isReturn = true;
-                });
-            });
-            if (isReturn) return pokemon;
-        });
-        handleFilterBestPokemon(bestPokemon);
-        // const newArr = (popularPokemon ?? []).filter((pokemon) => {
-        //     let isReturn = false;
-        //     pokemon.Types.forEach((type) => {
-        //         bestDamage.forEach((best) => {
-        //             if (best.zhName === type.zhName) isReturn = true;
-        //         });
-        //     });
-        //     if (isReturn) return pokemon;
-        // });
-        // handleFilterBestPokemon(newArr);
-    }, [bestDamage]);
-};
